requirements.txt
transformers>=4.30.0
torch             
accelerate
huggingface-hub
pillow
streamlit
python-dotenv